requirements.txt
img2pdf==0.4.4
lxml==4.8.0
MouseInfo==0.1.3
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
pikepdf==5.1.1
Pillow==9.1.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.7
pynput==1.7.6
pyparsing==3.0.8
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
six==1.16.0
